requirements.txt
click
rich
pyyaml
tinydb
requests
aiohttp
jinja2
